added helper methods, added a check that the playlist isn't already in the group's pool
