+++ back-end/helper_methods/playlist_is_in_pool.js
@@ -1,0 +1,38 @@
+let Playlist = require("../models/playlists.model");
+
+const playlist_is_in_pool = async (playlist_id, group_id) =>
+{
+  let pool
+  let passed = await Playlist.findOne({_id:group_id}) //retrieve the playlist with the given group id
+    .then(response => {
+      pool = response.pool
+      return true //no error encountered, so error will be set to null
+    })
+    .catch(err => {
+      const msg = "Error: Could not find group with given group id" 
+      console.log(msg)
+      console.log(err)
+      return new Error(msg)
+    })
+
+  if (passed instanceof Error)
+  {
+    return passed
+  }
+
+  for (let i = 0; i < pool.length; i++) 
+  { 
+    const user_playlist = pool[i]
+    const cur_playlist_id = user_playlist["playlist_id"]
+    if (cur_playlist_id === playlist_id)
+    {
+      return true
+    }
+  }
+
+  return false
+}
+
+module.exports = {
+  playlist_is_in_pool: playlist_is_in_pool
+}
