+++ back-end/requests/put/add_to_pool.js
@@ -1,63 +1,9 @@
 const axios = require("axios")
 let Playlist = require("../../models/playlists.model");
 const set_authentication = require("../other/authentication.js").set_authentication
-
-const is_valid_playlist = async (bearer, playlist_id) => {
-  if (!set_authentication(bearer, axios)) //set authentication
-  {   
-    console.log("Invalid authentication, could not validate playlist!")
-    return false
-  }
-
-let playlist_exists = await axios(`https://api.spotify.com/v1/playlists/${playlist_id}`) //make a request for the playlist
-  .then(response => { //if there is a valid response, the playlist must exist
-    console.log("Playlist validated for add_to_pool method")
-    return true
-  })
-  .catch(err => { //if an error is returned, the playlist doesn't exist (or something else was wrong with request)
-    console.log("add_to_pool error: Not a valid playlist")
-    console.error(err)
-    return false
-  })
-
-  return playlist_exists
-}
-
-const is_in_group = async (user_id, group_id) => { //checks if the user is a member of the specified group, and is also not banned
-  let members
-  let banned_members
-  let passed = await Playlist.findOne({_id:group_id}) //retrieve the playlist with the given group id
-    .then(response => {
-      members = response.members
-      banned_members = response.banned_members
-      return true //no error encountered, so error will be set to null
-    })
-    .catch(err => {
-      const msg = "Error: Could not find group with given group id" 
-      console.log(msg)
-      console.log(err)
-      return new Error(msg)
-    })
-
-  if (passed instanceof Error)
-  {
-    return passed
-  }
-
-  if (!(members.includes(user_id))) //user is not in the members list
-  {
-    const msg = "Error: member not in member list"
-    console.log(msg)
-    return new Error(msg)
-  } else if (banned_members.includes(user_id))
-  {
-    const msg = "Error: member is banned"
-    console.log(msg)
-    return new Error(msg)
-  }
-
-  return true
-}
+const is_in_group = require("../../helper_methods/is_in_group.js").is_in_group
+const is_valid_playlist =  require("../../helper_methods/is_valid_playlist.js").is_valid_playlist
+const playlist_is_in_pool = require("../../helper_methods/playlist_is_in_pool").playlist_is_in_pool
 
 const add_to_pool = async (req, res, next) => {
   const group_id = req.params.group_id
@@ -83,11 +29,27 @@ const add_to_pool = async (req, res, next) => {
     return next(in_group)
   }
 
+  //check that given playlist isn't already in the pool
+  let in_pool = await (playlist_is_in_pool(playlist_id, group_id))
+  
+  if (in_pool instanceof Error)
+  {
+    console.log("Error encountered in playlist_is_in_pool within add_to_pool.js")
+    return next(in_pool)
+  } else if (in_pool)
+  {
+    const msg = "Error: playlist to be added already is in the group's pool"
+    console.log(msg)
+    return next(new Error(msg))
+  }
+
   const playlist = {
     "added_by": user_id,
     "playlist_id": playlist_id,
   }
 
+
+  //check if the playlist is already in the pool
   await Playlist.updateOne({_id:group_id},
   {
     $push:{pool:playlist}
@@ -95,6 +57,7 @@ const add_to_pool = async (req, res, next) => {
   {safe: true, upsert: true}
   )
 
+  console.log("Playlist successfully added to pool")
   res.send("successfully added playlist to pool!")
 }
 module.exports = {
